requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-google-genai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search